File icons for macOS and Linux. Closes #565.

+++ j-lawyer-client/src/com/jdimension/jlawyer/client/utils/FileUtils.java
@@ -672,6 +672,7 @@ import java.io.FileOutputStream;
 import java.text.SimpleDateFormat;
 import java.util.Hashtable;
 import javax.swing.Icon;
+import javax.swing.ImageIcon;
 import javax.swing.JDialog;
 import javax.swing.JOptionPane;
 import javax.swing.JRootPane;
@@ -721,16 +722,64 @@ public class FileUtils extends ServerFileUtils {
         }
 
         try {
-            //Create a temporary file with the specified extension
-            File file = File.createTempFile("icon", fileExt);
 
-            FileSystemView view = FileSystemView.getFileSystemView();
-            Icon icon = view.getSystemIcon(file);
-            this.iconCache.put(fileExt, icon);
+            String osName = System.getProperty("os.name").toLowerCase();
 
-            //Delete the temporary file
-            file.delete();
-            return icon;
+            if (osName.indexOf("linux") > -1) {
+
+                if(fileExt.startsWith("."))
+                    fileExt=fileExt.substring(1,fileExt.length());
+                fileExt=fileExt.toLowerCase();
+                ImageIcon image = null;
+                try {
+                    image=new ImageIcon(getClass().getResource("/icons16/fileicons/file_type_" + fileExt + ".png"));
+                } catch (Throwable t) {
+                    log.warn("no file type icon for " + fileExt);
+                }
+                if (image != null) {
+                    this.iconCache.put(fileExt, image);
+                    return image;
+                } else {
+                    //Create a temporary file with the specified extension
+                    File file = File.createTempFile("icon", fileExt);
+
+                    FileSystemView view = FileSystemView.getFileSystemView();
+                    Icon icon = view.getSystemIcon(file);
+                    this.iconCache.put(fileExt, icon);
+
+                    //Delete the temporary file
+                    file.delete();
+                    return icon;
+                }
+            } else if (osName.startsWith("mac")) {
+
+                //Create a temporary file with the specified extension
+                File file = File.createTempFile("icon", fileExt);
+
+                FileSystemView view = FileSystemView.getFileSystemView();
+                Icon icon = FileSystemView.getFileSystemView().getSystemIcon(file);
+                if (icon != null) {
+                    this.iconCache.put(fileExt, icon);
+                }
+
+                //Delete the temporary file
+                file.delete();
+                return icon;
+
+            } else {
+                // Windows behaviour is default
+                //Create a temporary file with the specified extension
+                File file = File.createTempFile("icon", fileExt);
+
+                FileSystemView view = FileSystemView.getFileSystemView();
+                Icon icon = view.getSystemIcon(file);
+                this.iconCache.put(fileExt, icon);
+
+                //Delete the temporary file
+                file.delete();
+                return icon;
+            }
+
         } catch (Throwable t) {
             log.error("Could not determine default file type icon for " + fileName, t);
             return null;
@@ -746,7 +795,7 @@ public class FileUtils extends ServerFileUtils {
         return getNewFileName(currentFileName, datetimePrefix, d, EditorsRegistry.getInstance().getMainWindow(), "Datei umbenennen");
 
     }
-    
+
     public static String getNewFileName(String currentFileName, boolean datetimePrefix, java.util.Date d, Component parent) {
 
         return getNewFileName(currentFileName, datetimePrefix, d, parent, "Datei umbenennen");
@@ -801,7 +850,7 @@ public class FileUtils extends ServerFileUtils {
         return newFileName;
 
     }
-    
+
     public static String createTempFile(String fileName, byte[] content) throws Exception {
         return createTempFile(fileName, content, false);
     }
@@ -818,7 +867,7 @@ public class FileUtils extends ServerFileUtils {
         FileOutputStream fos = new FileOutputStream(new File(tmpFile), false);
         fos.write(content);
         fos.close();
-        
+
         if (readOnly) {
             try {
                 new File(tmpFile).setReadOnly();
